// frontend/packages/workflow/components/src/workflow-modal/type.ts
export enum MineActiveEnum {
  All = '1',
  Mine = '2',
}

/** data source */
export enum DataSourceType {
  /** process */
  Workflow = 'workflow',
  /** @deprecated process store */
  Product = 'product',
}

export type WorkflowItemType =
  | { type: DataSourceType.Workflow; item: WorkflowInfo }
  | { type: DataSourceType.Product; item: ProductInfo };

export const WORKFLOW_LIST_STATUS_ALL = 'all';
/**
 * In the category of workflows within a project when subprocesses are added, the resource library/project workflow category
 */
export enum WorkflowCategory {
  /**
   * project workflow
   */
  Project = 'project',
  /**
   * repository workflow
   */
  Library = 'library',
  /**
   * official example
   */
  Example = 'example',
}
/** Process pop-up status */
export interface WorkflowModalState {
  /** process state */
  status: WorkFlowListStatus | typeof WORKFLOW_LIST_STATUS_ALL;
  /** @Deprecated data type, is the process data or store data being requested? */
  dataSourceType: DataSourceType;
  /** creator */
  creator: MineActiveEnum;
  /** @deprecated workflow template tag */
  workflowTag: number;
  /** @deprecated product label */
  productCategory: string;
  /** Search keywords */
  query: string;
  /** @Deprecated whether to request the current space flow */
  isSpaceWorkflow: boolean;
  /** Selected workflow category */
  workflowCategory?: WorkflowCategory;
  /** Sort by @deprecated store products  */
  sortType?: SortType;
  /** The workflow type for list filtering in the pop-up window, the possible values are All, Workflow, Chatflow. Used for workflow type filtering in the list, at this time Imageflow has been merged into the Workflow type */
  listFlowMode: WorkflowMode;
}